Add round increment/decrement feature.
Known bug: Rapidly changing states causes weird-looking glitches
until the most recent update is received back from the server.
The right fix is probably to assume we're right and not accept changes
from the server for things we initiated. I think that means the
server wants to broadcast to everyone BUT the one who initiated each
update.